requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
get==0.0.39
idna==2.6
jsonschema==2.6.0
mmh3==2.5.1
optimizely-sdk==1.4.0
post==0.0.26
public==0.0.65
pycparser==2.18
pyOpenSSL==17.5.0
query-string==0.0.28
request==0.0.26
requests==2.18.4
six==1.11.0
urllib3==1.22
